+++ include/openPMD/IO/ADIOS/ADIOS2Auxiliary.hpp
@@ -56,12 +56,36 @@ namespace adios_defs
         Disk_Override
     };
 
+    using FlushTarget = adios_defs::FlushTarget;
+    FlushTarget flushTargetFromString(std::string const &str);
+
     enum class UseGroupTable
     {
         Yes,
         No
     };
 } // namespace adios_defs
+
+/*
+ * The following strings are used during parsing of the JSON configuration
+ * string for the ADIOS2 backend.
+ */
+namespace ADIOS2Defaults
+{
+    using const_str = char const *const;
+    constexpr const_str str_engine = "engine";
+    constexpr const_str str_type = "type";
+    constexpr const_str str_params = "parameters";
+    constexpr const_str str_usesteps = "usesteps";
+    constexpr const_str str_flushtarget = "preferred_flush_target";
+    constexpr const_str str_usesstepsAttribute = "__openPMD_internal/useSteps";
+    constexpr const_str str_adios2Schema =
+        "__openPMD_internal/openPMD2_adios2_schema";
+    constexpr const_str str_isBoolean = "__is_boolean__";
+    constexpr const_str str_activeTablePrefix = "__openPMD_groups";
+    constexpr const_str str_groupBasedWarning =
+        "__openPMD_internal/warning_bugprone_groupbased_encoding";
+} // namespace ADIOS2Defaults
 
 #if openPMD_HAVE_ADIOS2
 namespace detail
