+++ src/IO/ADIOS/ADIOS2Auxiliary.cpp
@@ -28,6 +28,37 @@
 
 #include <iostream>
 
+namespace openPMD::adios_defs
+{
+FlushTarget flushTargetFromString(std::string const &str)
+{
+    if (str == "buffer")
+    {
+        return FlushTarget::Buffer;
+    }
+    else if (str == "disk")
+    {
+        return FlushTarget::Disk;
+    }
+    else if (str == "buffer_override")
+    {
+        return FlushTarget::Buffer_Override;
+    }
+    else if (str == "disk_override")
+    {
+        return FlushTarget::Disk_Override;
+    }
+    else
+    {
+        throw error::BackendConfigSchema(
+            {"adios2", "engine", ADIOS2Defaults::str_flushtarget},
+            "Flush target must be either 'disk' or 'buffer', but "
+            "was " +
+                str + ".");
+    }
+}
+} // namespace openPMD::adios_defs
+
 namespace openPMD::detail
 {
 template <typename T>
